+++ docs/news.rst
@@ -14,6 +14,7 @@ Psycopg 3.3.0 (unreleased)
 ^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 - Drop support for Python 3.8.
+- Add support for binary hstore (:ticket:`#1030`).
 
 
 Psycopg 3.2.7 (unreleased)